+++ packages/qiankun/src/apis/isRuntimeCompatible.ts
@@ -1,0 +1,1 @@
+export { isRuntimeCompatible } from '@qiankunjs/shared';