+++ packages/qiankun/src/core/loadApp.ts
@@ -2,10 +2,11 @@
  * @author Kuitos
  * @since 2023-04-25
  */
-import type { ImportOpts } from '@qiankunjs/loader';
+import type { LoaderOpts } from '@qiankunjs/loader';
 import { loadEntry } from '@qiankunjs/loader';
 import type { Sandbox } from '@qiankunjs/sandbox';
 import { createSandboxContainer } from '@qiankunjs/sandbox';
+import { moduleResolver as defaultModuleResolver, transpileAssets, wrapFetchWithCache } from '@qiankunjs/shared';
 import { concat, isFunction, mergeWith } from 'lodash';
 import type { ParcelConfigObject } from 'single-spa';
 import getAddOns from '../addons';
@@ -18,6 +19,7 @@ import {
   performanceMeasure,
   toArray,
 } from '../utils';
+import { version } from '../version';
 
 export type ParcelConfigObjectGetter = (remountContainer: HTMLElement) => ParcelConfigObject;
 
@@ -25,9 +27,21 @@ export default async function loadApp<T extends ObjectType>(
   app: LoadableApp<T>,
   configuration?: AppConfiguration,
   lifeCycles?: LifeCycles<T>,
-) {
+): Promise<ParcelConfigObjectGetter> {
   const { name: appName, entry, container } = app;
-  const { fetch = window.fetch, sandbox, globalContext = window } = configuration || {};
+  const defaultNodeTransformer: AppConfiguration['nodeTransformer'] = (node, opts) => {
+    const moduleResolver = (url: string) => defaultModuleResolver(url, microAppDOMContainer, document.head);
+    return transpileAssets(node, entry, { ...opts, moduleResolver });
+  };
+  const {
+    fetch = window.fetch,
+    sandbox = true,
+    globalContext = window,
+    nodeTransformer = defaultNodeTransformer,
+    ...restConfiguration
+  } = configuration || {};
+
+  const fetchWithLruCache = wrapFetchWithCache(fetch);
 
   const markName = `[qiankun] App ${appName} Loading`;
   if (process.env.NODE_ENV === 'development') {
@@ -35,25 +49,36 @@ export default async function loadApp<T extends ObjectType>(
   }
 
   let global = globalContext;
-  let mountSandbox = () => Promise.resolve();
+  let mountSandbox: (container: HTMLElement) => Promise<void> = () => Promise.resolve();
   let unmountSandbox = () => Promise.resolve();
   let sandboxInstance: Sandbox | undefined;
 
-  let microAppContainer: HTMLElement = container;
-  clearContainer(microAppContainer);
+  let microAppDOMContainer: HTMLElement = container;
+  initContainer(microAppDOMContainer, appName, sandbox);
 
   if (sandbox) {
-    const sandboxContainer = createSandboxContainer(appName, () => microAppContainer, {
+    const sandboxContainer = createSandboxContainer(appName, () => microAppDOMContainer, {
       globalContext,
       extraGlobals: {},
+      fetch: fetchWithLruCache,
+      nodeTransformer,
     });
 
     sandboxInstance = sandboxContainer.instance;
     global = sandboxInstance.globalThis;
 
-    mountSandbox = () => sandboxContainer.mount();
+    mountSandbox = (domContainer) => sandboxContainer.mount(domContainer);
     unmountSandbox = () => sandboxContainer.unmount();
   }
+
+  const containerOpts: LoaderOpts = {
+    fetch: fetchWithLruCache,
+    sandbox: sandboxInstance,
+    nodeTransformer,
+    ...restConfiguration,
+  };
+
+  const lifecyclesPromise = loadEntry<MicroAppLifeCycles>(entry, microAppDOMContainer, containerOpts);
 
   const assetPublicPath = calcPublicPath(entry);
   const {
@@ -65,15 +90,11 @@ export default async function loadApp<T extends ObjectType>(
   } = mergeWith({}, getAddOns(global, assetPublicPath), lifeCycles, (v1, v2) =>
     concat((v1 ?? []) as LifeCycleFn<T>, (v2 ?? []) as LifeCycleFn<T>),
   );
-
+  // FIXME Due to the asynchronous execution of loadEntry, the DOM of the sub-app is inserted synchronously through appendChild, and inline scripts are also executed synchronously. Therefore, the beforeLoad may need to rely on transformer configuration to coordinate and ensure the order of asynchronous operations.
   await execHooksChain(toArray(beforeLoad), app, global);
 
-  const containerOpts: ImportOpts = { fetch, sandbox: sandboxInstance };
-
-  const lifecycles = await loadEntry<MicroAppLifeCycles>(entry, microAppContainer, containerOpts);
-
+  const lifecycles = await lifecyclesPromise;
   if (!lifecycles) throw new QiankunError(`${appName} entry ${entry} load failed as it not export lifecycles`);
-
   const { bootstrap, mount, unmount, update } = getLifecyclesFromExports(
     lifecycles,
     appName,
@@ -81,9 +102,9 @@ export default async function loadApp<T extends ObjectType>(
     sandboxInstance?.latestSetProp,
   );
 
-  const parcelConfigGetter: ParcelConfigObjectGetter = (remountContainer) => {
-    microAppContainer = remountContainer;
+  let mountTimes = 1;
 
+  return (mountContainer) => {
     const parcelConfig: ParcelConfigObject = {
       name: appName,
 
@@ -102,16 +123,22 @@ export default async function loadApp<T extends ObjectType>(
           }
         },
         async () => {
-          // if micro app container has no children that means now is remounting, we need to rerender the app manually
-          if (microAppContainer.firstChild === null) {
-            const htmlString = await getPureHTMLStringWithoutScripts(entry, fetch);
-            await loadEntry(htmlString, microAppContainer, containerOpts);
+          microAppDOMContainer = mountContainer;
+
+          // while the micro app is remounting, we need to load the entry manually
+          if (mountTimes > 1) {
+            initContainer(mountContainer, appName, sandbox);
+            // html scripts should be removed to avoid repeatedly execute
+            const htmlString = await getPureHTMLStringWithoutScripts(entry, fetchWithLruCache);
+            await loadEntry(htmlString, mountContainer, containerOpts);
           }
         },
-        mountSandbox,
+        async () => {
+          await mountSandbox(mountContainer);
+        },
         // exec the chain after rendering to keep the behavior with beforeLoad
         async () => execHooksChain(toArray(beforeMount), app, global),
-        async (props) => mount({ ...props, container: microAppContainer }),
+        async (props) => mount({ ...props, container: mountContainer }),
         // finish loading after app mounted
         async () => execHooksChain(toArray(afterMount), app, global),
         async () => {
@@ -120,19 +147,18 @@ export default async function loadApp<T extends ObjectType>(
             performanceMeasure(measureName, markName);
           }
         },
+        async () => {
+          mountTimes++;
+        },
       ],
 
       unmount: [
         async () => execHooksChain(toArray(beforeUnmount), app, global),
-        async (props) => unmount({ ...props, container: microAppContainer }),
+        async (props) => unmount({ ...props, container: mountContainer }),
         unmountSandbox,
         async () => execHooksChain(toArray(afterUnmount), app, global),
         async () => {
-          clearContainer(microAppContainer);
-          // for gc
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          microAppContainer = null;
+          clearContainer(mountContainer);
         },
       ],
     };
@@ -145,15 +171,19 @@ export default async function loadApp<T extends ObjectType>(
 
     return parcelConfig;
   };
-
-  return parcelConfigGetter;
 }
 
-function clearContainer(container: HTMLElement) {
-  if (!container) {
-    throw new QiankunError('container is not existed');
+function initContainer(container: HTMLElement, appName: string, sandboxCfg: AppConfiguration['sandbox']): void {
+  while (container.firstChild) {
+    container.removeChild(container.firstChild);
   }
 
+  container.dataset.name = appName;
+  container.dataset.version = version;
+  container.dataset.sandboxCfg = JSON.stringify(sandboxCfg);
+}
+
+function clearContainer(container: HTMLElement): void {
   while (container.firstChild) {
     container.removeChild(container.firstChild);
   }