+++ packages/sandbox/src/__test__/index.test.ts
@@ -1,7 +1,0 @@
-// import { greet } from '../index';
-//
-// test('should greet with name', () => {
-//   const name = 'qiankun3';
-//
-//   expect(greet(name)).toBe('Hello qiankun3');
-// });