+++ packages/sandbox/src/patchers/dynamicAppend/forStandardSandbox.ts
@@ -1,19 +1,25 @@
+/* eslint-disable */
 /**
  * @author Kuitos
  * @since 2020-10-13
  */
 
-import { nativeDocument, nativeGlobal } from '../../consts';
+import { Deferred, QiankunError } from '@qiankunjs/shared';
+import type { noop } from 'lodash';
+import { nativeDocument, nativeGlobal, qiankunHeadTagName } from '../../consts';
+import { rebindTarget2Fn } from '../../core/membrane/utils';
 import type { Sandbox } from '../../core/sandbox';
-import { isBoundedFunction, isCallable } from '../../utils';
 import type { Free } from '../types';
 import {
   calcAppCount,
+  getContainerBodyElement,
   getContainerHeadElement,
+  getNewRemoveChild,
+  getOverwrittenAppendChildOrInsertBefore,
   isAllAppsUnmounted,
-  patchHTMLDynamicAppendPrototypeFunctions,
   rebuildCSSRules,
   recordStyledComponentsCSSRules,
+  styleElementRefNodeNo,
   styleElementTargetSymbol,
 } from './common';
 import type { SandboxConfig } from './types';
@@ -25,8 +31,7 @@ declare global {
   }
 
   interface Window {
-    __sandboxConfigWeakMap__: WeakMap<Sandbox, SandboxConfig>;
-    __currentLockingSandbox__: Sandbox;
+    __currentLockingSandbox__?: Sandbox;
   }
 
   interface Document {
@@ -43,95 +48,231 @@ Object.defineProperty(nativeGlobal, '__currentLockingSandbox__', {
   configurable: true,
 });
 
-// eslint-disable-next-line @typescript-eslint/unbound-method
-const rawHeadAppendChild = HTMLHeadElement.prototype.appendChild;
-
-// Share sandboxConfigWeakMap between multiple qiankun instance, thus they could access the same record
-nativeGlobal.__sandboxConfigWeakMap__ = nativeGlobal.__sandboxConfigWeakMap__ || new WeakMap<Sandbox, SandboxConfig>();
-const sandboxConfigWeakMap = nativeGlobal.__sandboxConfigWeakMap__;
+const sandboxConfigWeakMap = new WeakMap<Sandbox, SandboxConfig>();
 
 const elementAttachSandboxConfigMap = new WeakMap<HTMLElement, SandboxConfig>();
-const mutationObserverPatchedMap = new WeakMap<
-  typeof MutationObserver.prototype.observe,
-  typeof MutationObserver.prototype.observe
->();
-const parentNodePatchedMap = new WeakMap<PropertyDescriptor, PropertyDescriptor>();
+const patchCacheWeakMap = new WeakMap<object, unknown>();
 
-function patchDocument(sandbox: Sandbox) {
+const getSandboxConfig = (element: HTMLElement) => elementAttachSandboxConfigMap.get(element);
+
+function patchDocument(sandbox: Sandbox, getContainer: () => HTMLElement): CallableFunction {
+  const container = getContainer();
+  // dom container might be reused by multiple apps,
+  // thus we check its attached sandbox is same with current to avoid duplicate patch
+  if (patchCacheWeakMap.get(container) === sandbox) {
+    return () => {};
+  }
+
+  const unpatch = patchDocumentHeadAndBodyMethods(container);
+
   const attachElementToSandbox = (element: HTMLElement) => {
     const sandboxConfig = sandboxConfigWeakMap.get(sandbox);
     if (sandboxConfig) {
       elementAttachSandboxConfigMap.set(element, sandboxConfig);
     }
   };
-
+  const getDocumentHeadElement = () => {
+    const container = getContainer();
+    const containerHeadElement = getContainerHeadElement(container);
+    if (!containerHeadElement) {
+      throw new QiankunError(`${sandbox.name} head element not existed while accessing document.head!`);
+    }
+    return containerHeadElement;
+  };
+  const getDocumentBodyElement = () => {
+    const container = getContainer();
+    return getContainerBodyElement(container);
+  };
+  const modificationFns: {
+    createElement?: typeof document.createElement;
+    querySelector?: typeof document.querySelector;
+  } = {};
   const proxyDocument = new Proxy(document, {
+    /**
+     * Read and write must be paired, otherwise the write operation will leak to the global
+     */
     set: (target, p, value) => {
-      target[p as string] = value;
+      switch (p) {
+        case 'createElement': {
+          modificationFns.createElement = value;
+          break;
+        }
+        case 'querySelector': {
+          modificationFns.querySelector = value;
+          break;
+        }
+        default:
+          target[p as keyof Document] = value;
+          break;
+      }
+
       return true;
     },
     get: (target, p, receiver) => {
-      if (p === 'createElement') {
-        // Must store the original createElement function to avoid error in nested sandbox
-        // eslint-disable-next-line @typescript-eslint/unbound-method
-        const targetCreateElement = target.createElement;
-        return function createElement(...args: Parameters<typeof document.createElement>) {
-          if (!nativeGlobal.__currentLockingSandbox__) {
-            nativeGlobal.__currentLockingSandbox__ = sandbox;
-          }
+      switch (p) {
+        case 'createElement': {
+          // Must store the original createElement function to avoid error in nested sandbox
+          const targetCreateElement = modificationFns.createElement || target.createElement;
+          return function createElement(...args: Parameters<typeof document.createElement>) {
+            if (!nativeGlobal.__currentLockingSandbox__) {
+              nativeGlobal.__currentLockingSandbox__ = sandbox;
+            }
 
-          const element = targetCreateElement.call(target, ...args);
+            const element = targetCreateElement.call(target, ...args);
 
-          // only record the element which is created by the current sandbox, thus we can avoid the element created by nested sandboxes
-          if (nativeGlobal.__currentLockingSandbox__ === sandbox) {
-            attachElementToSandbox(element);
-            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-            // @ts-ignore
-            delete nativeGlobal.__currentLockingSandbox__;
-          }
+            // only record the element which is created by the current sandbox, thus we can avoid the element created by nested sandboxes
+            if (nativeGlobal.__currentLockingSandbox__ === sandbox) {
+              attachElementToSandbox(element);
+              delete nativeGlobal.__currentLockingSandbox__;
+            }
 
-          return element;
-        };
+            return element;
+          };
+        }
+
+        case 'querySelector': {
+          const targetQuerySelector = modificationFns.querySelector || target.querySelector;
+          return function querySelector(...args: Parameters<typeof document.querySelector>) {
+            const selector = args[0];
+            switch (selector) {
+              case 'head': {
+                return getDocumentHeadElement();
+              }
+
+              case 'body': {
+                return getDocumentBodyElement();
+              }
+            }
+
+            return targetQuerySelector.call(target, ...args);
+          };
+        }
+
+        case 'head': {
+          return getDocumentHeadElement();
+        }
+
+        case 'body': {
+          return getDocumentBodyElement();
+        }
+
+        default:
+          break;
       }
 
-      const value = target[p as string];
+      const value = target[p as keyof Document];
       // must rebind the function to the target otherwise it will cause illegal invocation error
-      if (isCallable(value) && !isBoundedFunction(value)) {
-        return function proxyFunction(...args: unknown[]): unknown {
-          return Function.prototype.apply.call(value, target, args.map((arg) => (arg === receiver ? target : arg)));
-        };
-      }
-
-      return value;
+      return rebindTarget2Fn(target, value, receiver);
     },
   });
 
-  sandbox.addIntrinsics({ document: { value: proxyDocument, writable: false, enumerable: true, configurable: true } });
+  sandbox.addIntrinsics({
+    document: { value: proxyDocument, writable: false, enumerable: true, configurable: true },
+  });
 
+  patchCacheWeakMap.set(container, sandbox);
+
+  return () => {
+    unpatch();
+  };
+}
+
+function patchDocumentHeadAndBodyMethods(container: HTMLElement): typeof noop {
+  const patchHeadElementMethod = (headElement: HTMLHeadElement) => {
+    headElement.appendChild = getOverwrittenAppendChildOrInsertBefore(
+      document.head.appendChild,
+      getSandboxConfig,
+      'head',
+    );
+    headElement.insertBefore = getOverwrittenAppendChildOrInsertBefore(
+      document.head.insertBefore,
+      getSandboxConfig,
+      'head',
+    );
+    headElement.removeChild = getNewRemoveChild(document.head.removeChild, getSandboxConfig);
+  };
+  let containerHeadElement = getContainerHeadElement(container);
+  if (!containerHeadElement) {
+    // patch container head element after it is mounted
+    const observer = new MutationObserver(() => {
+      containerHeadElement = getContainerHeadElement(container);
+      if (containerHeadElement) {
+        patchHeadElementMethod(containerHeadElement);
+        observer.disconnect();
+      }
+    });
+    observer.observe(container, { subtree: true, childList: true });
+  } else {
+    patchHeadElementMethod(containerHeadElement);
+  }
+
+  const containerBodyElement = container;
+  containerBodyElement.appendChild = getOverwrittenAppendChildOrInsertBefore(
+    document.body.appendChild,
+    getSandboxConfig,
+    'body',
+  );
+  containerBodyElement.insertBefore = getOverwrittenAppendChildOrInsertBefore(
+    document.head.insertBefore,
+    getSandboxConfig,
+    'body',
+  );
+  containerBodyElement.removeChild = getNewRemoveChild(document.body.removeChild, getSandboxConfig);
+
+  return () => {
+    if (containerHeadElement) {
+      // @ts-ignore
+      delete containerHeadElement.appendChild;
+      // @ts-ignore
+      delete containerHeadElement.insertBefore;
+      // @ts-ignore
+      delete containerHeadElement.removeChild;
+    }
+
+    // @ts-ignore
+    delete containerBodyElement.appendChild;
+    // @ts-ignore
+    delete containerBodyElement.insertBefore;
+    // @ts-ignore
+    delete containerBodyElement.removeChild;
+  };
+}
+
+function patchDOMPrototypeFns(): typeof noop {
   // patch MutationObserver.prototype.observe to avoid type error
   // https://github.com/umijs/qiankun/issues/2406
-  // eslint-disable-next-line @typescript-eslint/unbound-method
   const nativeMutationObserverObserveFn = MutationObserver.prototype.observe;
-  if (!mutationObserverPatchedMap.has(nativeMutationObserverObserveFn)) {
-    const observe = function (this: MutationObserver, target: Node, options: MutationObserverInit) {
+  if (!patchCacheWeakMap.has(nativeMutationObserverObserveFn)) {
+    const observe = function observe(this: MutationObserver, target: Node, options: MutationObserverInit) {
       const realTarget = target instanceof Document ? nativeDocument : target;
       return nativeMutationObserverObserveFn.call(this, realTarget, options);
     };
 
     MutationObserver.prototype.observe = observe;
-    mutationObserverPatchedMap.set(nativeMutationObserverObserveFn, observe);
+    patchCacheWeakMap.set(nativeMutationObserverObserveFn, observe);
   }
 
+  // patch Node.prototype.compareDocumentPosition to avoid type error
+  const prevCompareDocumentPosition = Node.prototype.compareDocumentPosition;
+  if (!patchCacheWeakMap.has(prevCompareDocumentPosition)) {
+    Node.prototype.compareDocumentPosition = function compareDocumentPosition(this: Node, node) {
+      const realNode = node instanceof Document ? nativeDocument : node;
+      return prevCompareDocumentPosition.call(this, realNode);
+    };
+    patchCacheWeakMap.set(prevCompareDocumentPosition, Node.prototype.compareDocumentPosition);
+  }
+
+  // TODO https://github.com/umijs/qiankun/pull/2415 Not support yet as getCurrentRunningApp api is not reliable
   // patch parentNode getter to avoid document === html.parentNode
   // https://github.com/umijs/qiankun/issues/2408#issuecomment-1446229105
-  const parentNodeDescriptor = Object.getOwnPropertyDescriptor(Node.prototype, 'parentNode');
-  // if (parentNodeDescriptor && !parentNodePatchedMap.has(parentNodeDescriptor)) {
+  // const parentNodeDescriptor = Object.getOwnPropertyDescriptor(Node.prototype, 'parentNode');
+  // if (parentNodeDescriptor && !patchCacheWeakMap.has(parentNodeDescriptor)) {
   //   const { get: parentNodeGetter, configurable } = parentNodeDescriptor;
   //   if (parentNodeGetter && configurable) {
   //     const patchedParentNodeDescriptor = {
   //       ...parentNodeDescriptor,
   //       get(this: Node) {
-  //         const parentNode = parentNodeGetter.call(this);
+  //         const parentNode = parentNodeGetter.call(this) as HTMLElement;
   //         if (parentNode instanceof Document) {
   //           const proxy = getCurrentRunningApp()?.window;
   //           if (proxy) {
@@ -144,20 +285,27 @@ function patchDocument(sandbox: Sandbox) {
   //     };
   //     Object.defineProperty(Node.prototype, 'parentNode', patchedParentNodeDescriptor);
   //
-  //     parentNodePatchedMap.set(parentNodeDescriptor, patchedParentNodeDescriptor);
+  //     patchCacheWeakMap.set(parentNodeDescriptor, patchedParentNodeDescriptor);
   //   }
   // }
 
   return () => {
     MutationObserver.prototype.observe = nativeMutationObserverObserveFn;
-    mutationObserverPatchedMap.delete(nativeMutationObserverObserveFn);
+    patchCacheWeakMap.delete(nativeMutationObserverObserveFn);
 
-    if (parentNodeDescriptor) {
-      Object.defineProperty(Node.prototype, 'parentNode', parentNodeDescriptor);
-      parentNodePatchedMap.delete(parentNodeDescriptor);
-    }
+    Node.prototype.compareDocumentPosition = prevCompareDocumentPosition;
+    patchCacheWeakMap.delete(prevCompareDocumentPosition);
+
+    // if (parentNodeDescriptor) {
+    //   Object.defineProperty(Node.prototype, 'parentNode', parentNodeDescriptor);
+    //   patchCacheWeakMap.delete(parentNodeDescriptor);
+    // }
   };
 }
+
+// FIXME should not use global variable, should get it every time it is used, otherwise it may miss the runtime container or the business itself monkey patch logic
+const rawHeadInsertBefore = HTMLHeadElement.prototype.insertBefore;
+const rawHeadAppendChild = HTMLHeadElement.prototype.appendChild;
 
 export function patchStandardSandbox(
   appName: string,
@@ -165,28 +313,26 @@ export function patchStandardSandbox(
   opts: {
     sandbox: Sandbox;
     mounting?: boolean;
-  },
+  } & Pick<SandboxConfig, 'fetch' | 'nodeTransformer'>,
 ): Free {
-  const { sandbox, mounting = true } = opts;
+  const { sandbox, mounting = true, nodeTransformer, fetch } = opts;
   let sandboxConfig = sandboxConfigWeakMap.get(sandbox);
   if (!sandboxConfig) {
     sandboxConfig = {
       appName,
       sandbox,
-      getContainer: getContainer,
+      fetch,
+      nodeTransformer,
       dynamicStyleSheetElements: [],
+      dynamicExternalSyncScriptDeferredList: [],
     };
     sandboxConfigWeakMap.set(sandbox, sandboxConfig);
   }
   // all dynamic style sheets are stored in proxy container
   const { dynamicStyleSheetElements } = sandboxConfig;
 
-  const unpatchDynamicAppendPrototypeFunctions = patchHTMLDynamicAppendPrototypeFunctions(
-    (element) => elementAttachSandboxConfigMap.has(element),
-    (element) => elementAttachSandboxConfigMap.get(element)!,
-  );
-
-  const unpatchDocument = patchDocument(sandbox);
+  const unpatchDocument = patchDocument(sandbox, getContainer);
+  const unpatchDOMPrototype = patchDOMPrototypeFns();
 
   if (!mounting) calcAppCount(appName, 'increase', 'bootstrapping');
   if (mounting) calcAppCount(appName, 'increase', 'mounting');
@@ -195,29 +341,71 @@ export function patchStandardSandbox(
     if (!mounting) calcAppCount(appName, 'decrease', 'bootstrapping');
     if (mounting) calcAppCount(appName, 'decrease', 'mounting');
 
+    // release the overwritten document
+    unpatchDocument();
+
     // release the overwritten prototype after all the micro apps unmounted
     if (isAllAppsUnmounted()) {
-      unpatchDynamicAppendPrototypeFunctions();
-      unpatchDocument();
+      unpatchDOMPrototype();
     }
 
     recordStyledComponentsCSSRules(dynamicStyleSheetElements as HTMLStyleElement[]);
 
     // As now the sub app content all wrapped with a special id container,
-    // the dynamic style sheet would be removed automatically while unmounting
-
-    return function rebuild() {
-      rebuildCSSRules(dynamicStyleSheetElements as HTMLStyleElement[], (stylesheetElement) => {
-        const appWrapper = getContainer();
-        if (!appWrapper.contains(stylesheetElement)) {
-          const mountDom =
-            stylesheetElement[styleElementTargetSymbol] === 'head' ? getContainerHeadElement(appWrapper) : appWrapper;
-          rawHeadAppendChild.call(mountDom, stylesheetElement);
-          return true;
-        }
-
+    // the dynamic style sheet could be removed automatically while unmounting
+    return async function rebuild(container: HTMLElement) {
+      const isElementExisted = (element: HTMLStyleElement | HTMLLinkElement) => {
+        if (container.contains(element)) return true;
+        if ('rel' in element && element.rel === 'stylesheet' && element.href)
+          return !!container.querySelector(`link[rel=stylesheet][href=${element.href}]`);
         return false;
-      });
+      };
+
+      await Promise.all(
+        rebuildCSSRules(dynamicStyleSheetElements, async (stylesheetElement) => {
+          if (!isElementExisted(stylesheetElement)) {
+            const mountDom =
+              stylesheetElement[styleElementTargetSymbol] === 'head'
+                ? (() => {
+                    const containerHeadElement = getContainerHeadElement(container);
+                    if (!containerHeadElement) {
+                      throw new QiankunError(
+                        `${appName} container ${qiankunHeadTagName} element not ready while rebuilding!`,
+                      );
+                    }
+                    return containerHeadElement;
+                  })()
+                : container;
+
+            let styleElement = stylesheetElement;
+
+            const deferred = new Deferred<boolean>();
+            if ('rel' in styleElement && styleElement.rel === 'stylesheet' && styleElement.href) {
+              // micro app rendering should wait unit the rebuilding link element is loaded, otherwise it may cause style blink
+              // As one external link element will just trigger loaded event once, although we append it multiple times, we need to clone it before every appending
+              styleElement = styleElement.cloneNode(true) as HTMLLinkElement;
+              styleElement.onload = () => deferred.resolve(true);
+              styleElement.onerror = () => deferred.resolve(false);
+            } else {
+              deferred.resolve(true);
+            }
+
+            const refNo = stylesheetElement[styleElementRefNodeNo];
+            if (typeof refNo === 'number' && refNo !== -1) {
+              // the reference node may be dynamic script comment which is not rebuilt while remounting thus reference node no longer exists
+              // in this case, we should append the style element to the end of mountDom
+              const refNode = mountDom.childNodes[refNo];
+              rawHeadInsertBefore.call(mountDom, styleElement, refNode);
+            } else {
+              rawHeadAppendChild.call(mountDom, styleElement);
+            }
+
+            return deferred.promise;
+          }
+
+          return false;
+        }),
+      );
     };
   };
 }